Add script to generate README from json file
Having the same information in two places is quite
painful and proves to lead to a few errors here and
there.
The generate file is not exactly similar to the
README I've commited because there are a few
discrepencies that I do not quite know how to handle
(related to multiple links for slides/other resources)
as I did not dare changing the format of the json file
not to break things.